+++ dao/src/main/java/com/getjavajob/training/yarginy/socialnetwork/dao/relations/manytomany/selfrelated/SelfManyToManyDml.java
@@ -9,10 +9,10 @@ import java.sql.SQLException;
 import java.util.Collection;
 
 public abstract class SelfManyToManyDml<E extends Entity> {
-    public abstract Collection<E> select(Connection connection, int id) throws SQLException;
+    public abstract Collection<E> select(Connection connection, long id) throws SQLException;
 
-    public abstract PreparedStatement getSelectStatement(Connection connection, int firstId, int secondId) throws
+    public abstract PreparedStatement getSelectStatement(Connection connection, long firstId, long secondId) throws
             SQLException;
 
-    public abstract void updateRow(ResultSet resultSet, int firstId, int secondId) throws SQLException;
+    public abstract void updateRow(ResultSet resultSet, long firstId, long secondId) throws SQLException;
 }
